Polymorphism examples
If you are an avid crafter, you may know multiple different crafts in which
you could create a fabric. And depending on the type or style of fabric you
want, you would use a different craft. But at the end of the day, you want
to do the same thing: create fabric. If we were to model that programmatically,
we would take advantage of polymorphism.

Polymorphism is an object oriented concept where you can use multiple classes
in exactly the same way so that their concrete class doesn't matter. Inheritance
is one way to achieve polymorphism. But other ways can include duck typing
(defining the same method signature on multiple classes) or using interfaces
(if you are using a statically typed language that supports them).

In this code example, we use duck typing to achieve polymorphism where each
yarn craft class defines a `createFabric` method, allowing us to use them all
exactly the same way without caring about which concrete class we're using at
that moment.