Set Theory examples
Venn diagrams are a great way to think about set theory. A set is a data
structure similar to arrays or lists but it is an *unordered collection*
of objects with *no duplicates*.

If you have 2 Venn diagrams where the circle on the left is X and the
circle on the right is Y, basic set theory can be described as follows.

Union (set notation: X ∪ Y) = all elements that exist in X OR Y (all the things!)

Intersection (set notation: X ∩ Y) = all elements that exist in X AND Y
(only the stuff in the middle)

Difference (set notation: X - Y) = only the elements that exist in X and
not in the middle

Relative Complement (set notation: Y \ X) = the same as Y - X, only the
elements that exist in Y and not in the middle

Symmetric Difference (set notation: X △ Y) = the elements that exist in only X
and the elements that exist in only Y but none of the stuff in the middle
This is also known as the disjunctive union and is the same as (X ∖ Y) ∪ (Y ∖ X)
(if you want to be fancy)