Recursion examples
Russian nesting dolls are a great metaphor for recursion, because each doll is
the same except for its size. The dolls continue to open until you get to
the smallest child which does not open. When you reach the smallest child,
your reverse the process, closing each doll one by one in reverse order.

Recursion is the process by which a function calls itself directly or indirectly.
The function that is calling itself is called the "recursive function". Everything
that you can do recursively, you can also do iteratively so you can think of
recursion as fancy looping.

To ensure that you don't continue calling the recursive function infinitely,
causing a stack overflow exception, you have to set up a condition where it
exits the continued nesting and returns a finite value. This is called the
*base case*. The smallest doll that does not open in a Russian nesting doll set
is the base case.