Linked list examples
When you are following a scavenger hunt, you start with the first clue and must
follow each clue sequentially one at a time. Because you have no way of knowing
where the nth clue is when you start, you must follow the hunt from clue to clue
to clue.

A scavenger hunt is a good metaphor for a linked list. A linked list is a data
structure characterized by sequential data access and no random access. This
is unlike arrays. In arrays, because all of the data is stored in contiguous
locations in memory, you can do simple addition and subtraction to find the memory
location of the data at the nth node. For example, array[3] is just 3 memory
locations from array[0]. Linked lists on the other hand, do not need contiguous
memory allocated. Each node in a linked list points to where the next node is located.

Because of the memory allocation, a linked list will have no wasted space
because it can grow and shrink dynamically whereas traditionally an array always needs to be
allocated to maximum size (or copied to a new array if it needs to grow).
We tend to forget this when working with Javascript since JS Arrays behave
more like ArrayList.

However, because a linked list node needs to store its data and a reference
to the next node, it does take up a bit more space than an array of the same
data.

Linked lists have fast insertion at the tail and insertion/deletion at the head. But
to insert into the middle of a linked list is O(n) because we need to *loop
through the list* to the place we want to insert, update the previous nodes pointer
to the inserted node and update the inserted nodes pointer to the next node.