Big O Notation examples
Cooking with ratios is when you don't memorize recipes but instead memorize
the ratios of ingredients. The amount of ingredients you need changes
_in proportion_ to how much of the food you want to make. For example,
cupcakes follow a 4:3:2:1 ratio.

Big O Notation measures the running time (or complexity, memory usage, and
other resources) _in proportion_ to the size of a method or function's input.

The actual input size is unimportant because we want to measure the proportional
complexity of the logic.

O(1) = constant time. As the input size changes, the time to execute remains the same.

O(n) = directly proportional to the input size. As the input size increases,
the time to execute increases linearly. This is common in loops.

O(n^2) = proportional to the square of the input size. This is common in nested loops.

O(2^n) = time to execute grows exponentially with the size of the input. For example,
calculating Fibonacci numbers recursively is O(2^n).

O(logn) = this can kind of be thought of as opposite of 2^n. A common example
of O(logn) is binary search where every iteration loops through half as many
items as the iteration prior. So rather than the input doubling in size every iteration
(like O(2^n)), the input halves in size every iteration.